merge from Pcdc dev for version 0.1.3

+++ Jenkinsfile
@@ -3,10 +3,14 @@
 // See 'Loading libraries dynamically' here: https://jenkins.io/doc/book/pipeline/shared-libraries/
 library 'cdis-jenkins-lib@master'
 
+import org.jenkinsci.plugins.pipeline.modeldefinition.Utils
+
 node {
   def AVAILABLE_NAMESPACES = ['jenkins-blood', 'jenkins-brain', 'jenkins-niaid', 'jenkins-dcp', 'jenkins-genomel']
   List<String> namespaces = []
   List<String> listOfSelectedTests = []
+  skipUnitTests = false
+  skipQuayImgBuildWait = false
   doNotRunTests = false
   kubectlNamespace = null
   kubeLocks = []
@@ -22,10 +26,8 @@ node {
     stage('FetchCode'){
       gitHelper.fetchAllRepos(pipeConfig['currentRepoName'])
     }
-    stage('gen3 helper test suite') {
-      sh 'GEN3_HOME=$WORKSPACE/cloud-automation XDG_DATA_HOME=$WORKSPACE/dataHome bash cloud-automation/gen3/bin/testsuite.sh --profile jenkins'
-    }
     stage('CheckPRLabels') {
+     try {
       // giving a chance for auto-label gh actions to catch up
       // sleep(10)
       for(label in prLabels) {
@@ -38,9 +40,19 @@ node {
             selectedTest = "suites/" + selectedTestLabel[1] + "/" + selectedTestLabel[2] + ".js"
             listOfSelectedTests.add(selectedTest)
             break
+          case "skip-gen3-helper-tests":
+            println('Skipping unit tests assuming they have been verified in a previous PR check iteration...')
+            skipUnitTests = true
+            break
+          case "skip-awshelper-build-wait":
+            println('Skipping the WaitForQuayBuild stage as it is not necessary for every PR...')
+            skipQuayImgBuildWait = true
+            break
           case "doc-only":
             println('Skip tests if git diff matches expected criteria')
             doNotRunTests = docOnlyHelper.checkTestSkippingCriteria()
+            skipUnitTests = true
+            skipQuayImgBuildWait = true
             break
           case "debug":
             println("Call npm test with --debug")
@@ -72,17 +84,42 @@ node {
       if (listOfSelectedTests.isEmpty()) {
 	  listOfSelectedTests.add("all")
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)  
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
+    }
+    stage('gen3 helper test suite') {
+     try {
+      println("namespaces: ${namespaces}")
+      if(!skipUnitTests) {
+        sh 'GEN3_HOME=$WORKSPACE/cloud-automation XDG_DATA_HOME=$WORKSPACE/dataHome bash cloud-automation/gen3/bin/testsuite.sh --profile jenkins'
+      } else {
+        Utils.markStageSkippedForConditional(STAGE_NAME)
+      }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('gen3 helper test suite with zsh') {
-      if(!doNotRunTests) {
+     try {
+      if(!skipUnitTests) {
         sh 'GEN3_HOME=$WORKSPACE/cloud-automation XDG_DATA_HOME=$WORKSPACE/dataHome zsh cloud-automation/gen3/bin/testsuite.sh --profile jenkins'
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
-
     stage('pytest') {
-      if(!doNotRunTests) {
+     try {
+      if(!skipUnitTests) {
         sh 'pip3 install boto3 --upgrade'
         sh 'pip3 install kubernetes --upgrade'
         sh 'python -m pytest cloud-automation/apis_configs/'
@@ -94,36 +131,60 @@ node {
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('nginx helper test suite') {
-      if(!doNotRunTests) {
+     try {
+      if(!skipUnitTests) {
         dir('cloud-automation/kube/services/revproxy') {
           sh 'npx jasmine helpersTest.js'
         }
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('python 2 base image dockerrun.sh test') {
-      if(!doNotRunTests) {
+     try {
+      if(!skipUnitTests) {
         dir('cloud-automation/Docker/python-nginx/python2.7-alpine3.7') {
           sh 'sh dockerrun.sh --dryrun=True'
         }
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('python 3 base image dockerrun.sh test') {
-      if(!doNotRunTests) {
+     try {
+      if(!skipUnitTests) {
         dir('cloud-automation/Docker/python-nginx/python3.6-alpine3.7') {
           sh 'sh dockerrun.sh --dryrun=True'
         }
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('WaitForQuayBuild') {
-      if(!doNotRunTests) {
+     try {
+      if(!skipQuayImgBuildWait) {
         quayHelper.waitForBuild(
           "awshelper",
           pipeConfig['currentBranchFormatted']
@@ -131,16 +192,28 @@ node {
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('SelectNamespace') {
-     if(!doNotRunTests) {
+     try {
+      if(!doNotRunTests) {
         (kubectlNamespace, lock) = kubeHelper.selectAndLockNamespace(pipeConfig['UID'], namespaces)
         kubeLocks << lock
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('ModifyManifest') {
+     try {
       if(!doNotRunTests) {
         manifestHelper.editService(
           kubeHelper.getHostname(kubectlNamespace),
@@ -150,9 +223,15 @@ node {
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     
     stage('K8sReset') {
+     try {
       if(!doNotRunTests) {
         // adding the reset-lock lock in case reset fails before unlocking
         kubeLocks << kubeHelper.newKubeLock(kubectlNamespace, "gen3-reset", "reset-lock")
@@ -160,23 +239,41 @@ node {
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('VerifyClusterHealth') {
+     try {
       if(!doNotRunTests) {
         kubeHelper.waitForPods(kubectlNamespace)
         testHelper.checkPodHealth(kubectlNamespace, "")
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('GenerateData') {
+     try {
       if(!doNotRunTests) {    
         testHelper.simulateData(kubectlNamespace)
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('FetchDataClient') {
+     try {
       if(!doNotRunTests) {
         // we get the data client from master, unless the service being
         // tested is the data client itself, in which case we get the
@@ -189,8 +286,14 @@ node {
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('RunTests') {
+     try {
       if(!doNotRunTests) {
         testHelper.runIntegrationTests(
             kubectlNamespace,
@@ -202,15 +305,27 @@ node {
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
       }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
     }
     stage('CleanS3') {
-      if(!doNotRunTests) {    
-        testHelper.cleanS3()
+     try {
+      if(!doNotRunTests) {
+        testHelper.cleanS3(kubectlNamespace)
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
-      }    
+      }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true) 
     }
     stage('authzTest') {
+     try {
       if(!doNotRunTests) {
         // test revproxy+arborist /gen3-authz stuff
         kubeHelper.kube(kubectlNamespace, {
@@ -218,21 +333,29 @@ node {
         });
       } else {
         Utils.markStageSkippedForConditional(STAGE_NAME)
-      }   
-    }
+      }
+     } catch (ex) {
+        metricsHelper.writeMetricWithResult(STAGE_NAME, false)
+        throw ex
+     }
+     metricsHelper.writeMetricWithResult(STAGE_NAME, true)
+   }
   }
   catch (e) {
     pipelineHelper.handleError(e)
+    throw e
   }
   finally {
     stage('Post') {
       kubeHelper.teardown(kubeLocks)
       testHelper.teardown()
-      // tear down network policies deployed by the tests
-      kubeHelper.kube(kubectlNamespace, {
+      if(!skipUnitTests) {
+        // tear down network policies deployed by the tests
+        kubeHelper.kube(kubectlNamespace, {
           sh(script: 'kubectl --namespace="' + kubectlNamespace + '" delete networkpolicies --all', returnStatus: true);
         });
-      pipelineHelper.teardown(currentBuild.result)
+        pipelineHelper.teardown(currentBuild.result)
+      }
     }
   }
 }
